Separated tests for testserver
