+++ lib/testserver.js
@@ -1,17 +1,8 @@
-'use strict';
-
-/**
- * Test server to load test.
- * (C) 2013 Alex Fernández.
- */
-
-
-const testing = require('testing');
-const http = require('http');
-const WebSocketServer = require('websocket').server;
-const util = require('util');
-const net = require('net');
-const {Latency} = require('./latency.js');
+import * as http from 'http'
+import {server as WebSocketServer} from 'websocket'
+import * as util from 'util'
+import * as net from 'net'
+import {Latency} from './latency.js'
 
 const PORT = 7357;
 const LOG_HEADERS_INTERVAL_SECONDS = 1;
@@ -177,33 +168,8 @@ class TestServer {
  *	- error: set an HTTP error code, default is 500.
  * An optional callback is called after the server has started.
  */
-exports.startServer = function(options, callback) {
+export function startServer(options, callback) {
 	const server = new TestServer(options);
 	return server.start(callback);
-};
-
-function testStartServer(callback) {
-	const options = {
-		port: 10530,
-	};
-	const server = exports.startServer(options, error => {
-		testing.check(error, 'Could not start server', callback);
-		server.close(error => {
-			testing.check(error, 'Could not stop server', callback);
-			testing.success(callback);
-		});
-	});
-}
-
-/**
- * Run the tests.
- */
-exports.test = function(callback) {
-	testing.run([testStartServer], 5000, callback);
-};
-
-// start server if invoked directly
-if (__filename == process.argv[1]) {
-	exports.test(testing.show);
 }
